cloud_backend/requirements.txt
flask
gunicorn
google-genai
pillow